0.9.5_menus
-consolidated menu screens into single class
-removed separate menu classes

menu options are still created with the game and change display based on
current menu screen, as opposed to being created with the screen
