--- TicTacToe/src/game/screens/VersusMenu.java
@@ -1,50 +1,0 @@
-/*******************************************************************************
- * File Name:			
- * Project:			
- * 
- * Designer(s):		Garrett Cross,
- * 					Omar Kermiche,
- * 					Autumn Nguyen,
- * 					Thomas Pridy
- * 
- * Copyright © 2019. All rights reserved.
- ******************************************************************************/
-package game.screens;
-
-import game.TicTacToe;
-
-public class VersusMenu extends MenuScreen
-{   
-    /***************************************************************************
-     *      VARIABLES
-     **************************************************************************/
-
-    private int i = 0;
-    
-    /***************************************************************************
-     *      CONSTRUCTOR
-     **************************************************************************/
-    
-    public VersusMenu(TicTacToe game)
-    {
-        super(game);
-        setMenuType(Menu.VERSUS);
-        setMenuTitle();
-        setLabel(i++, "Computer");
-        setLabel(i++, "Human");
-        setLabel(i++, "Back");
-    }
-    
-    /***************************************************************************
-     *      SETTERS/GETTERS
-     **************************************************************************/
-
-    /***************************************************************************
-     *      METHODS
-     **************************************************************************/
-
-    /***************************************************************************
-     *      ENUMERATORS
-     **************************************************************************/
-
-}
